requirements.txt
aiogram
PyOpenGL
moderngl
Pillow
imageio
numpy
requests
